ToolbarWidget v1.1
Rewritten module.

Previous method (moving an element) caused the iframe to reload,
which broke content scripts and such.

The new method presets the position of the widget. The sdk/widget
module automatically recognizes this preference, and puts the
button in the expected place.

Look at sdk/widget.js and search for "_insertNodeInToolbar:"
 (sdk/widget.js , lines 597-652 in Jetpack 1.14)